requirements.txt
streamlit
pillow
torch
torchvision
transformers
numpy
opencv-python
scikit-learn
